honglou RAG/requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
pyvis
networkx
tqdm
